% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_mt_cov.R
\name{module_mito_server}
\alias{module_mito_server}
\title{Module mito Server}
\usage{
module_mito_server(id)
}
\arguments{
\item{id}{}
}
\description{
Module mito Server
}
\examples{
Need to add examples
TO DO FIll out this skeleton

}
