% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_volcano_ui.R
\name{module_volcano_ui}
\alias{module_volcano_ui}
\title{Module Volcano ui}
\usage{
module_volcano_ui(id)
}
\arguments{
\item{id}{parameter}
}
\value{
volcano plot
}
\description{
Module Volcano ui
}
\examples{
module_volcano_ui
TO DO need examples here
}
